requirements.txt
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
pytest==6.0.1
pytest-cov==2.10.1
pytest-django==3.9.0
pytest-mock==3.3.1